Separate out other class
